+++ Prototype/src/math/polygon.cpp
@@ -155,11 +155,11 @@ void Polyshape::renderPolygon(int x, int y, Color color)
     SDL_SetRenderDrawColor(GraphicsEngine::getActiveRenderer(), color.r, color.g, color.b, color.a);
     for (int i = 0; i < vertices.size(); i++) {
         //Weird code that makes sure polygons are drawn contained within its dimensions
-        int a = (abs(vertices[i].x) >= (int)(width) ? vertices[i].x - 1 : vertices[i].x) + x;
-        int b = (abs(vertices[i].y) >= (int)(height) ? vertices[i].y - 1 : vertices[i].y) + y;
-        int c = (abs(vertices[(i + 1) % vertices.size()].x) >= (int)width ?
+        int a = (int)(abs(vertices[i].x) >= (int)(width) ? vertices[i].x - 1 : vertices[i].x) + x;
+        int b = (int)(abs(vertices[i].y) >= (int)(height) ? vertices[i].y - 1 : vertices[i].y) + y;
+        int c = (int)(abs(vertices[(i + 1) % vertices.size()].x) >= (int)width ?
                  vertices[(i + 1) % vertices.size()].x - 1 : vertices[(i + 1) % vertices.size()].x) + x;
-        int d = (abs(vertices[(i + 1) % vertices.size()].y) >= (int)height ?
+        int d = (int)(abs(vertices[(i + 1) % vertices.size()].y) >= (int)height ?
             vertices[(i + 1) % vertices.size()].y - 1 : vertices[(i + 1) % vertices.size()].y) + y;
         SDL_RenderDrawLine(GraphicsEngine::getActiveRenderer(), a, b, c, d);
     }
